+++ providers/kmpauth-google/src/androidMain/kotlin/com/mmk/kmpauth/google/GoogleLegacyAuthentication.kt
@@ -29,7 +29,8 @@ internal class GoogleLegacyAuthentication(
     override suspend fun signIn(
         filterByAuthorizedAccounts: Boolean,
         isAutoSelectEnabled: Boolean,
-        scopes: List<String>
+        scopes: List<String>,
+        requestAccessToken: Boolean,
     ): Result<GoogleUser> {
         val signInClient = getGoogleSignInClient(scopes = scopes).signInIntent
         activityResultState.isInProgress = true