+++ providers/kmpauth-google/src/jvmMain/kotlin/com/mmk/kmpauth/google/GoogleAuthUiProviderImpl.kt
@@ -30,7 +30,8 @@ internal class GoogleAuthUiProviderImpl(private val credentials: GoogleAuthCrede
     override suspend fun signIn(
         filterByAuthorizedAccounts: Boolean,
         isAutoSelectEnabled: Boolean,
-        scopes: List<String>
+        scopes: List<String>,
+        requestAccessToken: Boolean,
     ): Result<GoogleUser> {
         val redirectTarget = resolveRedirectTarget()
             ?: return Result.failure(